requirements.txt
python-dotenv==0.19.1
instagrapi==1.15.5
Pillow>=8.1.1
